Title auto-complete functionality added
When a url is added to the 'URL' field, the title can be auto-completed using
the minimal-metainspector package. The url added to the field is sent to the
server which then looks up the title and returns it or serves an error if one
occured to the client. The client then adds the title to the 'Title' field
automatically.

Error handling has also been implemented - if the server finds an error with
the url, the client then displays error messages and adds the 'is-danger'
class to several elements to warn the user that the url entered is invalid
and needs to be corrected.