Error handling of invalid URLs
These changes prevent the submission of invalid URLs to the server. If there was
an error getting the title of an invalid URL, the 'submit' button becomes
disabled until a valid title and URL are entered.

Once the 'submit' button has been pressed, the title and URL input fields become
disabled. This was implemented because users would need to click on the 'view
tanzakus' button after the initial submission but the input fields would still
be editable. This change shows shows users that once submitted, their addition
has been locked in.