Integration of a PostgreSQL database
These changes add a PostgreSQL database to this project. This is mainly to
have persistance of data when deployed on Heroku since writing to the file
system is not allowed/does not work.

With PostgreSQL integrated, the webapp will be able to retrieve and add data
without needing to worry about data loss if the Heroku dyno goes to sleep.