Implement infinite scrolling
After much thought between whether to implement infinite scrolling or
pagination, I went with infinite scrolling. This commit defines a new
function `loadMore()` which loads the designated number of tanzakus as
defined by the constant `LOAD_TANZAKUS_AMOUNT` (defaults to 5).

`loadMore()` is called when the user reaches the bottom of the document
when scrolling, causing more tanzakus to be displayed. Fixes #6.